Splittin abs utils into multiple files to not include abs specific includes which can brake other sources
