+++ metadata-ingestion/src/datahub/ingestion/source/azure/abs_folder_utils.py
@@ -1,6 +1,4 @@
 import logging
-import os
-import re
 from typing import Dict, Iterable, List, Optional
 
 from azure.storage.blob import BlobProperties
@@ -10,65 +8,8 @@ from datahub.ingestion.api.common import PipelineContext
 from datahub.ingestion.source.azure.azure_common import AzureConnectionConfig
 from datahub.metadata.schema_classes import GlobalTagsClass, TagAssociationClass
 
-ABS_PREFIXES_REGEX = re.compile(
-    r"(http[s]?://[a-z0-9]{3,24}\.blob\.core\.windows\.net/)"
-)
-
 logging.getLogger("py4j").setLevel(logging.ERROR)
 logger: logging.Logger = logging.getLogger(__name__)
-
-
-def is_abs_uri(uri: str) -> bool:
-    return bool(ABS_PREFIXES_REGEX.match(uri))
-
-
-def get_abs_prefix(abs_uri: str) -> Optional[str]:
-    result = re.search(ABS_PREFIXES_REGEX, abs_uri)
-    if result and result.groups():
-        return result.group(1)
-    return None
-
-
-def strip_abs_prefix(abs_uri: str) -> str:
-    # remove abs prefix https://<storage-account>.blob.core.windows.net
-    abs_prefix = get_abs_prefix(abs_uri)
-    if not abs_prefix:
-        raise ValueError(
-            f"Not an Azure Blob Storage URI. Must match the following regular expression: {str(ABS_PREFIXES_REGEX)}"
-        )
-    length_abs_prefix = len(abs_prefix)
-    return abs_uri[length_abs_prefix:]
-
-
-def make_abs_urn(abs_uri: str, env: str) -> str:
-    abs_name = strip_abs_prefix(abs_uri)
-
-    if abs_name.endswith("/"):
-        abs_name = abs_name[:-1]
-
-    name, extension = os.path.splitext(abs_name)
-
-    if extension != "":
-        extension = extension[1:]  # remove the dot
-        return f"urn:li:dataset:(urn:li:dataPlatform:abs,{name}_{extension},{env})"
-
-    return f"urn:li:dataset:(urn:li:dataPlatform:abs,{abs_name},{env})"
-
-
-def get_container_name(abs_uri: str) -> str:
-    if not is_abs_uri(abs_uri):
-        raise ValueError(
-            f"Not an Azure Blob Storage URI. Must match the following regular expression: {str(ABS_PREFIXES_REGEX)}"
-        )
-    return strip_abs_prefix(abs_uri).split("/")[0]
-
-
-def get_key_prefix(abs_uri: str) -> str:
-    if not is_abs_uri(abs_uri):
-        raise ValueError(
-            f"Not an Azure Blob Storage URI. Must match the following regular expression: {str(ABS_PREFIXES_REGEX)}"
-        )
-    return strip_abs_prefix(abs_uri).split("/", maxsplit=1)[1]
 
 
 def get_abs_properties(
@@ -280,7 +221,3 @@ def list_folders(
             this_dict[folder_name] = folder_name
 
         yield f"{folder_name}"
-
-
-def get_container_relative_path(abs_uri: str) -> str:
-    return "/".join(strip_abs_prefix(abs_uri).split("/")[1:])
